Latent Dim 4 and Beta equal to 0.005
After finalizing β = 0.005 for latent dimension = 4, we fine-tuned the model. Using all healthy patient data via 6-fold leave-one-out validation. In each iteration, the model was trained on 5 out of 6 healthy patients and validated on the remaining one. This process was repeated 6 times—each time holding out a different healthy patient for validation—resulting in 6 independently trained versions of the model.